+++ devtools/common/src/commonTest/kotlin/com/maximbircu/devtools/common/extensions/DevToolExtensionsKtTest.kt
@@ -6,8 +6,12 @@ import com.maximbircu.devtools.common.presentation.tools.group.GroupTool
 import com.maximbircu.devtools.common.presentation.tools.text.TextTool
 import com.maximbircu.devtools.common.presentation.tools.toggle.ToggleTool
 import com.maximbircu.devtools.common.utils.mockk
+import com.maximbircu.devtools.common.utils.returns
 import io.mockk.verify
+import kotlinx.serialization.json.json
 import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertFailsWith
 
 class DevToolExtensionsKtTest {
     @Test
@@ -21,5 +25,86 @@ class DevToolExtensionsKtTest {
         verify { listener("first-tool", tools.getValue("first-tool")) }
         verify { listener("second-tool", groupTool) }
         verify { listener("child-tool", groupTool.tools.getValue("child-tool")) }
+    }
+
+    @Test
+    fun `returns proper JSON config`() {
+        val tools = mapOf<String, DevTool<*>>(
+            "first-tool" to createTool<ToggleTool> { ::value returns true },
+            "second-tool" to GroupTool(
+                mapOf(
+                    "first-child-tool" to createTool<TextTool> { ::value returns "First value" },
+                    "second-child-tool" to createTool<TextTool> { ::value returns "Second value" }
+                )
+            )
+        )
+        val expectedConfigJsonObject = json {
+            "first-tool" to true
+            "second-tool" to json {
+                "first-child-tool" to "First value"
+                "second-child-tool" to "Second value"
+            }
+        }
+
+        val configJsonObject = tools.toJsonObjectOfValues()
+
+        assertEquals(expectedConfigJsonObject, configJsonObject)
+    }
+
+    @Test
+    fun `returns proper filtered JSON config`() {
+        val tools = mapOf<String, DevTool<*>>(
+            "first-tool" to createTool<ToggleTool> {
+                ::value returns true
+                ::isEnabled returns true
+            },
+            "second-tool" to GroupTool(
+                mapOf(
+                    "first-child-tool" to createTool<TextTool> {
+                        ::value returns "First value"
+                        ::isEnabled returns false
+                    },
+                    "second-child-tool" to createTool<TextTool> {
+                        ::value returns "Second value"
+                        ::isEnabled returns false
+                    }
+                )
+            ),
+            "third-tool" to GroupTool(
+                mapOf(
+                    "first-child-tool" to createTool<TextTool> {
+                        ::value returns "First value"
+                        ::isEnabled returns false
+                    },
+                    "second-child-tool" to createTool<TextTool> {
+                        ::value returns "Second value"
+                        ::isEnabled returns true
+                    },
+                    "third-child-tool" to createTool<TextTool> {
+                        ::value returns 3.4
+                        ::isEnabled returns true
+                    }
+                )
+            )
+        )
+        val expectedConfigJsonObject = json {
+            "first-tool" to true
+            "third-tool" to json {
+                "second-child-tool" to "Second value"
+                "third-child-tool" to 3.4
+            }
+        }
+
+        val configJsonObject = tools.toJsonObjectOfValues { it.isEnabled }
+
+        assertEquals(expectedConfigJsonObject, configJsonObject)
+    }
+
+    @Test
+    fun `throws exception when tries to convert to primitive json element a non primitive type`() {
+        val tools = mapOf<String, DevTool<*>>(
+            "first-tool" to createTool<ToggleTool> { ::value returns object {} }
+        )
+        assertFailsWith<IllegalArgumentException> { tools.toJsonObjectOfValues() }
     }
 }